cleaned up

+++ src/_difficultyBeatmaps.kt
@@ -17,6 +17,6 @@ data class _difficultyBeatmaps (
 	@SerializedName("_difficulty") val _difficulty : String,
 	@SerializedName("_difficultyRank") val _difficultyRank : Int,
 	@SerializedName("_beatmapFilename") val _beatmapFilename : String,
-	@SerializedName("_noteJumpMovementSpeed") val _noteJumpMovementSpeed : Int,
-	@SerializedName("_noteJumpStartBeatOffset") val _noteJumpStartBeatOffset : Int
+	@SerializedName("_noteJumpMovementSpeed") val _noteJumpMovementSpeed : Double,
+	@SerializedName("_noteJumpStartBeatOffset") val _noteJumpStartBeatOffset : Double
 )
